Allow to delete a message queue alongside with its related data
When deleting a message queue, all queue's related queues and data will be deleted from the system. If the given queue has an online consumer/producer, an error will be returned. For determining if a consumer/producer is online, an additional heartbeat check is performed, so that crushed consumers/producers would not block queue deletion.