Make MultiQueueProducer publish a message as a single atomic operation
This commit fixes a case where a message is being published when at the same time the destination message queue is going to be deleted. This could cause data inconsistency where, in the worst case scenario, we could have a pending queue, or a priority queue, without its parent queue.