Allow configuring which messages to store with extended parameters
Configuring acknowledged and dead-lettered messages storage was limited, allowing either to store all messages or none. This commit adds the possibility to choose which messages to store and enables to set a queue size and a retention time for the messages.

Additionally, the broker has been refactored and cleaned up.